connecting with firestore
