+++ lib/features/home/widgets/get_books.dart
@@ -1,19 +1,18 @@
-import '../core/models/book_model.dart';
+import 'package:ketabna/core/models/book_model.dart';
 
-class GetBooks
-{
+class GetBooks {
   static final List<BookModel> book = [
     BookModel(
-        nameAr: '',
-        picture:'assets/images/Fatherhood.jpg',
-        nameEn: 'Fatherhood',
-        authorName: 'Marcus Berkmann',
-        category: '',
-        ownerUid: '',
+      nameAr: '',
+      picture: 'assets/images/Fatherhood.jpg',
+      nameEn: 'Fatherhood',
+      authorName: 'Marcus Berkmann',
+      category: '',
+      ownerUid: '',
     ),
     BookModel(
       nameAr: '',
-      picture:'assets/images/Dissapearance of Emile Zola.jpg',
+      picture: 'assets/images/Dissapearance of Emile Zola.jpg',
       nameEn: 'The Dissapearance of Emila Zola',
       authorName: 'Michael Rosen',
       category: '',
@@ -21,7 +20,7 @@ class GetBooks
     ),
     BookModel(
       nameAr: '',
-      picture:'assets/images/How To Be The Best In Time And Space.jpg',
+      picture: 'assets/images/How To Be The Best In Time And Space.jpg',
       nameEn: 'The Time Travellers Handbook',
       authorName: 'Stride Lottie',
       category: '',
