+++ lib/temp/book_item.dart
